requirements.txt
selenium
pillow
pyautogui
argparse
pypandoc
